+++ packages/npm/oa42-generator/src/generators/constants/index.ts
@@ -1,0 +1,1 @@
+export * from "./operation-accept.js";
