+++ packages/npm/oa42-generator/src/generators/constants/operation-accept.ts
@@ -1,0 +1,20 @@
+import * as models from "../../models/index.js";
+import { itt } from "../../utils/index.js";
+import { getOperationAcceptConstName, getOperationAcceptTypeName } from "../names/index.js";
+
+export function* generateOperationAcceptConstant(operationModel: models.Operation) {
+  const operationAcceptTypeName = getOperationAcceptTypeName(operationModel);
+  const operationAcceptConstName = getOperationAcceptConstName(operationModel);
+
+  const operationAccepts = [
+    ...new Set(
+      operationModel.operationResults.flatMap((operationResultModel) =>
+        operationResultModel.bodies.map((bodyModel) => bodyModel.contentType),
+      ),
+    ),
+  ];
+
+  yield itt`
+    export const ${operationAcceptConstName}: ${operationAcceptTypeName}[] = ${JSON.stringify(operationAccepts)};
+  `;
+}
