+++ packages/npm/oa42-generator/src/generators/types/index.ts
@@ -1,4 +1,6 @@
+export * from "./authentication-handler.js";
 export * from "./client-credentials.js";
+export * from "./operation-accept.js";
 export * from "./operation-authentication.js";
 export * from "./operation-handler.js";
 export * from "./operation-incoming-request.js";
